Fix the basics
Now a features can be enabled and active.
Futher functionallity e.g specifying roles and users/entities is done by
including modules with more behavior.
Note that he base functionality takes precedence over added
functionality. This means that if a function is set to enabled, then it
will always be enabled regardless of any roles and entities. And if the
feature is not active, then is always disabled regardless of any roles
and entities.